Cleaned up dmft_tools/solver.py, moved get_n_orbitals to dmft_tools/common.py, trimmed comments and imports
